// src/app/model/UsersGet.ts
export class UsersGet{
    roles: `String`;
    cedula: any;
    usuario: `String`;
    passwordd: any;
    primer_NOMBRE: any;
    id_USUARIO: number;
    pregunta_SECRETA: any;
    tipo_IDENTIFICACION: `String`;
    segundo_APELLIDO: any;
    correo_PRINCIPAL: `String`;
    primer_APELLIDO: any;
    segundo_CORREO: any;
    segundo_NOMBRE: any;
    usuario_CREACION: `String`;
    fecha_CREACION: `String`;
    usuario_MODIFICACION: any;
    respuesta_SECRETA: any;
    maquina_CREACION: `String`;
    maquina_ACTUALIZA: any;
    fecha_MODIFICACION: `String`;
    usuario_ACTIVO: `String`;
}